gift/gift-ui/src/main/java/com/epam/esm/gift/controller/PurchaseController.java
package com.epam.esm.gift.controller;

import com.epam.esm.gift.model.Purchase;
import com.epam.esm.gift.service.PurchaseService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("purchase")
public class PurchaseController {
    private final PurchaseService purchaseService;

    @Autowired
    public PurchaseController(PurchaseService purchaseService) {
        this.purchaseService = purchaseService;
    }

    @GetMapping
    public List<Purchase> getAll(@RequestParam(value = "page", required = false, defaultValue = "1") int page) {
        return purchaseService.getAll(page);
    }
}
